Implement gatsby-background-image for optimized background image
gatsby-background-image is a React component which for background-images provides, what Gatsby’s own gatsby-image does for the rest of your images and even more.

It has all the advantages of gatsby-image, including the “blur-up” technique or a ”traced placeholder” SVG to show a preview of the image while it loads,
plus being usable as a container (no more hacks with extra wrappers)
plus being able to work with multiple stacked background images
plus being able to style with Tailwind CSS and suchlike Frameworks

All the glamour (and speed) of gatsby-image for your Background Images!